src/use-cases/get-user-profile.ts
import { User } from '@prisma/client';

import { ResourceNotFoundError } from './errors/resource-not-found.error';

import { UsersRepository } from '@/repositories/users.repository';

interface GetUserProfileRequest {
  userId: string;
}

interface GetUserProfileResponse {
  user: User;
}

export class GetUserProfile {
  constructor(private usersRepository: UsersRepository) {}
  async execute({
    userId,
  }: GetUserProfileRequest): Promise<GetUserProfileResponse> {
    const user = await this.usersRepository.findById(userId);

    if (!user) {
      throw new ResourceNotFoundError();
    }

    return { user };
  }
}
